Fix media controls.

+++ Source/WebCore/platform/graphics/MediaPlayer.h
@@ -272,12 +272,12 @@ public:
     bool hasAudio() const;
 
     //Tell client what rates we support
-    void playbackRatesSupported(float *rates, int count) {
+    void playbackRatesSupported(double *rates, int count) {
         m_rates.clear();
         for (int i = 0; i < count; i++)
             m_rates.append(rates[i]);
     }
-    Vector<float> getPlayRates() { return m_rates; }
+    Vector<double> getPlayRates() { return m_rates; }
 
     void setFrameView(FrameView* frameView) { m_frameView = frameView; }
     FrameView* frameView() { return m_frameView; }
@@ -514,7 +514,7 @@ private:
     bool m_privateBrowsing;
     bool m_shouldPrepareToRender;
     bool m_contentMIMETypeWasInferredFromExtension;
-    Vector<float> m_rates;
+    Vector<double> m_rates;
 
 #if ENABLE(PLUGIN_PROXY_FOR_VIDEO)
     WebMediaPlayerProxy* m_playerProxy;    // not owned or used, passed to m_private
